Добавляет исправления. Вроде бы ни чего не сломал
