+++ index.html
@@ -138,7 +138,7 @@
     <!-- Полноэкранный показ изображения -->
     <section class="big-picture  overlay  hidden">
       <h2 class="big-picture__title  visually-hidden">Просмотр фотографии</h2>
-      <div class="big-ficture__preview">
+      <div class="big-picture__preview">
 
         <!-- Просмотр изображения -->
         <div class="big-picture__img">
