+++ js/createSimilarPhoto.js
@@ -1,0 +1,1 @@
+export const createSimilarPhoto = newPhoto;
